MySQL replaced with MSSQL. Rewritten repositories

+++ src/L2dotNET/Network/serverpackets/ExPostItemList.cs
@@ -29,7 +29,7 @@ namespace L2dotNET.Network.serverpackets
                 WriteShort(0);
                 WriteShort(0);
 
-                WriteInt(item.Template.BodyPart);
+                WriteInt((int) item.Template.BodyPart);
                 WriteShort(item.Enchant);
                 WriteShort(0);
