+++ src/L2dotNET/Network/serverpackets/TradeStart.cs
@@ -34,7 +34,7 @@ namespace L2dotNET.Network.serverpackets
                 WriteShort(item.Template.Type2);
                 WriteShort(item.CustomType1);
 
-                WriteInt(item.Template.BodyPart);
+                WriteInt((int) item.Template.BodyPart);
                 WriteShort(item.Enchant);
                 WriteShort(item.CustomType2);
